Update html files for deleting & viewing accounts
Added a href in ledger.html so that each account has it owns link. Need
to add a controller to handle this. Additional information is now
displayed for each account in the delete drop down box, to make it
easier for the user to select the correct account to be deleted.

Added constraints to LedgerAccount to prevent creating accounts with
no name and/or number.