Add transaction controller
The transaction controller handles adding new transactions. New
transactions are added through a form on the account page. After
adding a new transaction the account page is reloaded to display the
new transaction. Fixed date parsing since the format was incorrect
and the wrong date was being stored in the database.

A hidden field was added to the account.html that is used to pass the
account id to the controller. Also added a step to the amount input to
make it clear that this is a currency field.