Add a way to delete transactions in view
Entries that are seen on the accounts page in the account.html now
have a button next to the entry's info that gives the user the option
to delete the corresponding transaction. We have to delete the entire
transaction in order to comply with double-entry bookkeeping, because
our books wouldn't be balanced if we deleted only one entry from a
transaction.