Add updating transactions to transaction service
I added a method to the service layer that is used to update a
transaction given its id and all field values that can be modified. Had
to add a getter and setter for the creditAccount and debitAccount fields
so that I could modify them before updating the transaction in the
database.