Add editing transactions capability
Transactions can now be edited. If a user clicks on the edit glyphicon
that is next to every entry in the account.html page, it takes the user
to a page where they can edit the transaction that the entry corresponds
to.
On this page if the user submits the form than the transaction gets
updated in the database, and the user goes back to the account page where
they can view the updated transaction.