Polish LedgerAccountService saveAccount method
Previously the saveAccount service method that creates a new
ledgerAccount before saving it to the database would call the DAO's
saveAccount. Now it calls the saveAccount method that it is overloading.
This ensures that any business logic that may be added to the overloaded
method is applied to this method as well.